API: Expense - Create

+++ rs/src/main/java/ua/ll7/slot7/ma/data/request/ExpenseCreateRequest.java
@@ -8,7 +8,7 @@ import ua.ll7.slot7.ma.data.generic.StringLongDG;
  * 30.12.14 17:26
  */
 public class ExpenseCreateRequest extends StringLongDG {
-  private double data3;
+  private float data3;
 
   @Override
   public String toString() {
@@ -17,11 +17,11 @@ public class ExpenseCreateRequest extends StringLongDG {
                          "} " + super.toString();
   }
 
-  public double getData3() {
+  public float getData3() {
     return data3;
   }
 
-  public void setData3(double data3) {
+  public void setData3(float data3) {
     this.data3 = data3;
   }
 }
